+++ tests/pagination/test_order_by.py
@@ -1,11 +1,11 @@
 import pytest
 from rest_framework import exceptions
 
-import winter
+from winter.pagination import order_by
 
 
 def test_with_default_not_in_allowed_fields():
     with pytest.raises(exceptions.ParseError) as exception:
-        winter.pagination.order_by(['id'], default_sort=('uid',))
+        order_by(['id'], default_sort=('uid',))
 
     assert exception.value.args == ('Fields do not allowed as order by fields: "uid"',)