+++ winter/core/json/__init__.py
@@ -1,1 +1,4 @@
+from .decoder import DecodeException
+from .decoder import json_decode
+from .decoder import json_decoder
 from .encoder import JSONEncoder